+++ modules/title-bar-main-process.js
@@ -5,7 +5,7 @@ define([
     "vs/code/electron-main/app",
     "vs/base/common/platform",
     "vs/platform/configuration/common/configuration",
-    "customize-ui/utils"
+    "vscode-gtk-ui/utils"
 ], function (module, require, electron, app, platform, configuration, utils) {
     'use strict';