+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-All notable changes to the "customize-ui" extension will be documented in this file.
+All notable changes to the "vscode-gtk-ui" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
@@ -140,7 +140,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [0.1.21] - 2019-10-09
 
-- Fix https://github.com/iocave/customize-ui/issues/8
+- Fix https://github.com/iocave/vscode-gtk-ui/issues/8
 
 ## [0.1.20] - 2019-10-03